chore(sandcastle): inline pipeline.ts into main.ts
pipeline.ts existed to host the per-issue flow (reconcile → sandbox →
implement → maybe-review → PR-open), but with parallelism and the
subset planner removed there's only one call site and one issue per
iteration. The "module" was just bookkeeping around 90 lines that the
orchestrator loop reads top-to-bottom anyway.

Inline runIssuePipeline, the IssueOutcome / PipelineContext types,
describeExclusion, isDocsOnly, gitOutput, and pathsTouchedByCommits
into main.ts. Replace the Promise.allSettled-style outcome handling
with a straight try/catch, since one issue per iteration only ever
produces one outcome.

main.ts is now ~165 lines and reads as a single linear flow — closer
to the reference Sandcastle setups, with the bits of judgement we
actually need (eligibility, docs-only branching, all-markdown review
skip, reconciliation-lite) sitting where the loop runs them.
